+++ responsetype/responsetype.go
@@ -1,9 +1,18 @@
 package responsetype
 
-type ResponseType string
+type Ais string
 
 const (
-	Position   ResponseType = "Position"
-	Aton                    = "Aton"
-	Staticdata              = "Staticdata"
+	Position   Ais = "Position"
+	Aton           = "Aton"
+	Staticdata     = "Staticdata"
+)
+
+type Combined string
+
+const (
+	FullJson      Combined = "FullJson"
+	SimpleJson             = "SimpleJson"
+	FullGeojson            = "FullGeojson"
+	SimpleGeojson          = "SimpleGeojson"
 )